+++ app/static/js/alerts.js
@@ -3,7 +3,7 @@ function showAlert(message, type = 'success') {
 
   // Create the alert HTML and append it
   alertContainer.insertAdjacentHTML('beforeend', `
-    <div class="alert alert-${type} alert-dismissible fade show" role="alert">
+    <div class="alert alert-${type} alert-dismissible fade show m-0" role="alert">
         ${message}
         <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
     </div>