+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add basic anti-brute force lockout protection to login page.
 - Add ability to toggle password visibility to all password forms.
 - Add ability to reset TOTP 2fa secret via web-lgsm.py script.
 - Add TOTP Two Factor Auth option for web panel login!